refactor(#10): move useFirstRender
